% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_oras.R
\name{install_oras}
\alias{install_oras}
\title{install the oras client}
\usage{
install_oras(
  os = system_os(),
  arch = system_arch(),
  version = "1.0.0",
  path = bin_path(),
  force = FALSE
)
}
\arguments{
\item{os}{operating system}

\item{arch}{architecture}

\item{version}{which version should be installed}

\item{path}{destination where binary is installed.}

\item{force}{install even if binary is already found.
Can be used to force upgrade.}
}
\value{
path to the minio binary (invisibly)
}
\description{
install the oras client
}
\details{
This function is just a convenience wrapper for prebuilt \code{oras}
binaries. Supports Windows, Mac, and Linux on both Intel/AMD (amd64) and ARM
architectures.
For details, see official oras docs for your operating system,
e.g. \url{https://oras.land/docs/installation}.

NOTE: If you want to install to other than the default location, OR if you
already have a oras client installed somewhere and want to use that,
simply set the option "oras.bin.dir", to the appropriate location of the
directory containing your "oras" binary, e.g.
\verb{options("oras.bin.dir" = "/usr/local/bin)}.  Note that this package
will not automatically use MINIO available on $PATH (to promote security
and portability in design).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
install_oras()
\dontshow{\}) # examplesIf}
}
