% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oras.R
\name{oras}
\alias{oras}
\title{oras}
\usage{
oras(command, ..., path = bin_path(), verbose = TRUE)
}
\arguments{
\item{command}{space-delimited text string of an mc command
(starting after the mc ...)}

\item{...}{additional arguments to \code{\link[processx:run]{processx::run()}}}

\item{path}{location where oras executable will be installed. By default will
use the OS-appropriate storage location.}

\item{verbose}{print output?}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
The oras client
}
\details{
This function forms the basis for all other available commands.
run \code{oras("-h")} for help.
}
