% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oras_helpers.R
\name{oras_blob_fetch}
\alias{oras_blob_fetch}
\title{Pull objects from remote registry by hash}
\usage{
oras_blob_fetch(name, output, flags = "", verbose = TRUE)
}
\arguments{
\item{name}{of registry and digest of blob, as \verb{<name>@<digest>}}

\item{output}{filename to write blob to on local disk}

\item{flags}{additional flags, see \code{oras("blob fetch -h")}}

\item{verbose}{show output messages? logical, TRUE}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
Pull objects from remote registry by hash
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

sha <- "450a97ba6b438c6ea5bdf2aaac7eab0ecbbf812b5ff74b56f62dcf0a0c7eb0e5"
oras_blob_fetch(paste0("localhost:5000/hello@sha256:", sha), "mtcars.csv")
\dontshow{\}) # examplesIf}
}
