% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oras_helpers.R
\name{oras_blob_push}
\alias{oras_blob_push}
\title{Push object blobs from remote registry by hash}
\usage{
oras_blob_push(name, file, flags = "", verbose = TRUE)
}
\arguments{
\item{name}{of registry and optinal digest of blob, as \verb{<name>@<digest>}}

\item{file}{local file to push}

\item{flags}{additional flags, see \code{oras("blob push -h")}}

\item{verbose}{show output messages? logical, TRUE}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
Push object blobs from remote registry by hash
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

oras_blob_push(paste0("localhost:5000/hello"), "mtcars.csv")
\dontshow{\}) # examplesIf}
}
