% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oras_helpers.R
\name{oras_cp}
\alias{oras_cp}
\title{copy artifacts from one target to another}
\usage{
oras_cp(from, to, flags = "", verbose = TRUE)
}
\arguments{
\item{from}{artifact address}

\item{to}{artifact destination}

\item{flags}{see \code{oras_cp("-h")} for list of flags}

\item{verbose}{show output messages? logical, TRUE}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
copy artifacts from one target to another
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
oras_cp("localhost:5000/hello:v1", "localhost:5001/hello:v1")
\dontshow{\}) # examplesIf}
}
