% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oras_helpers.R
\name{oras_logout}
\alias{oras_logout}
\title{Log out from a remote registry}
\usage{
oras_logout(registry, verbose = TRUE)
}
\arguments{
\item{registry}{Address of the remote}

\item{verbose}{show output messages? logical, TRUE}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
Log out from a remote registry
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
oras_logout()
\dontshow{\}) # examplesIf}
}
