% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oras_helpers.R
\name{oras_pull}
\alias{oras_pull}
\title{Pull files from remote registry}
\usage{
oras_pull(name, flags = "", verbose = TRUE)
}
\arguments{
\item{name}{name (with tag or digest optional) of registry to pull from}

\item{flags}{additional flags, see \code{oras_pull("-h")}}

\item{verbose}{show output messages? logical, TRUE}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
Pull files from remote registry
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
oras_pull("localhost:5000/hello:v1")
\dontshow{\}) # examplesIf}
}
