% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oras_helpers.R
\name{oras_push}
\alias{oras_push}
\title{Push files to remote registry}
\usage{
oras_push(name, file, flags = "", verbose = TRUE)
}
\arguments{
\item{name}{of file in registry (optionally with tag or hash)}

\item{file}{local file to push}

\item{flags}{additional flags, see \code{oras_push("-h")}}

\item{verbose}{show output messages? logical, TRUE}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
Push files to remote registry
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
oras_push("localhost:5000/hello:v1", "hi.txt")
\dontshow{\}) # examplesIf}
}
