% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oras_helpers.R
\name{oras_tag}
\alias{oras_tag}
\title{Tag a manifest in the remote registry}
\usage{
oras_tag(name, new_tag, flags, verbose = TRUE)
}
\arguments{
\item{name}{of file in registry (optionally with tag or hash)}

\item{new_tag}{additional tag to add}

\item{flags}{additional flags, see \code{oras("push -h")}}

\item{verbose}{show output messages? logical, TRUE}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
Tag a manifest in the remote registry
}
